# model/openapi.yaml
openapi: 3.0.3
info:
  title: eComm REST API
  version: 1.0.0
  description: |-
    An e-commerce application REST API that allows users to perform various CRUD
    operations such as registering an account, browsing products for sale, etc. 
    <br>
    ## Authentication & Authorization 
    This API uses cookies to authenticate client requests and maintain session information.
    <br>
    Most of the endpoints in this API require an authenticated session to exist.
    <br>
    Unfortunately, [Swagger UI doesn't support cookie authentication for "try it out" requests](https://swagger.io/docs/specification/authentication/cookie-authentication/) due to browser security restrictions.
    <br>
    This means that to test endpoints marked as requiring authentication, a tool like [Postman](https://www.postman.com/) or [REST Client](https://marketplace.visualstudio.com/items?itemName=humao.rest-client) would have to be used.
    <br>
    All other endpoints can be tested directly from this page using the "try it out" feature.
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: connect.sid
      in: cookie
paths:
  /register:
    summary: Register a new user
    post:
      summary: Registers a new user
      description: Creates a new user in the database with the supplied credentials, and creates a login session for the user
      operationId: register
      requestBody:
        description: Must inclcude `username` and `password`. *Optional:* `email`.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  description: Username
                  type: string
                password:
                  description: password
                  type: string
                email:
                  description: email address
                  type: string
              required:
                - username
                - password
            example:
              username: testUser
              password: 10102020aabb
              email: testUser@test.com
      responses:
        "201":
          description: Successfully registered user
        "409":
          description: Conflict! User already exists
        "500":
          description: Internal error
      tags:
        - Authentication
  /login:
    summary: Login
    post:
      summary: Logs in a registered user
      operationId: login
      requestBody:
        description: User login
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  description: Username
                  type: string
                password:
                  description: password
                  type: string
      responses:
        "200":
          description: Successfully logged in
        "400":
          description: Bad request
        "401":
          description: Unauthorized. Invalid username/password
      tags:
        - Authentication
    get:
      summary: Checks for an active user session
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Active user session found
        "404":
          description: No active session found
      tags:
        - Authentication
  /logout:
    summary: Logout
    post:
      summary: Logs out current logged-in user session
      security:
        - cookieAuth: []
      operationId: logout
      responses:
        "200":
          description: Successfully logged out
        "401":
          description: Unauthorized (no active user to log out)
      tags:
        - Authentication
...